Got Lexer Doing something
The lexer seems to work and can tokenize a simple Javascript-like
language. It does this using regex. It will currently ignore whitespace
and newlines (not tabs though, should add). An example file is in the
folder called `testfile.txt`. To run it using with the test file, does
`make test` or `make && ./bin/hhfi testfile.txt`. You can still do
`make run`, it will just give a single EOF token in the stream.