Greatly Reworked Lexer tokenization
The lexer tokenization (primarily the function) was greatly reworked to
make it significantly faster (I think most likely O(n) or close). It is
using a Trie in order to store the pattern information and matches until
it finds the biggest possible match.

The tokenization runs in an order of priority of:
 - Keywords
 - Literals
 - Operators
 - Separators
 - Identifiers/Variables

`main.cpp` currently runs through a string and tokenizes it. The easiest
way to test with a file is to run `make test`, although a manual running
with a command line argument would be fine.