Moved Lexeme data to Lexer
The data for matching the Lexemes was previously global. It is now moved
to the Lexer class to remove the global variable and add polymorphism.
In addition, a BasicLexer class was added that loads the data that was
previously global. This will allow for the lexers to be changed simply
by changing the type of the object created.